==> fast-detection-of-duplicate-bug-report/Step 1-Data_preprocessing.ipynb <==
# **Data Preprocessing**

import pandas as pd
import nltk
import gensim
from gensim.utils import simple_preprocess
from nltk.corpus import stopwords
from nltk import word_tokenize
from nltk.stem import WordNetLemmatizer, SnowballStemmer
from nltk.stem.porter import *
from gensim.parsing.preprocessing import STOPWORDS
import numpy as np
import pickle
nltk.download('stopwords')
nltk.download('wordnet')

import warnings
warnings.filterwarnings('ignore')

#mounting the google drive
#from google.colab import drive
#drive.mount('/content/drive')

#importing the CSV file of the dataset from google drive into a dataframe
data=pd.read_csv('dataset/eclipse_platform.csv')
data

#Removing the unuseful coloumns from the dataset
data = data.drop(columns=['Priority', 'Component', 'Status', 'Resolution', 'Version', 'Created_time', 'Resolved_time'], axis=1)
data

#Printing the Coloumn names
data.keys()

#Checking the shape of the data
data.shape


## **1. Text Cleaning**

  *   Removing Invalid Reports
  *   Removing Punctuations






#Number of empty Description reports
data['Description'].isnull().sum()

#Number of empty Title reports
data['Title'].isnull().sum()

#removing the rows with empty Description
data = data.dropna(axis=0, subset=['Description'])

pd.options.display.max_colwidth = 500

#Showing the fixed pattern used in many reports which doesn't have to do anything with the bug itself
# Eg: fixed in head, has been marked as read only
data.loc[[4]]

#Removing the fixed patterns by replacing them with " "
data["Description"]= data["Description"].str.replace("fixed in HEAD", "", case = False)
data["Description"]= data["Description"].str.replace("has been marked as readonly", " ", case = False)

#Checking if it is been removed
data.loc[[4]]

pd.options.display.max_colwidth = 50

#Text Cleaning round 1 (removing punctutions)
import re
import string

def clean_text_round1(text):
    '''Make text lowercase, remove text in square brackets, remove punctuation and remove words containing numbers.'''
    text = re.sub('\w*\d\w*', '', text)
    text = re.sub('\w*\f\w*', '', text)
    text = re.sub('\(.*?\)', '', text)
    text = re.sub('\[.*]\)', '', text)
    text = text.lower()
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)
    return text

round1 = lambda x: clean_text_round1(x)

data['Description'] = data['Description'].apply(clean_text_round1)


# Apply a second round of cleaning (removing punctuations)
def clean_text_round2(text):
    '''Get rid of some additional punctuation and non-sensical text that was missed the first time around.'''
    text = re.sub('[‘’“”…]', '', text)
    text = re.sub('\n', '', text)
    text = re.sub('\t', '', text)
    return text

round2 = lambda x: clean_text_round2(x)

data['Title'] = data['Title'].apply(clean_text_round2)
data['Description'] = data['Description'].apply(clean_text_round2)

## **2. Stop Word Removal, Tokenization and Lemmatization**


def lemmatize(text):
    return WordNetLemmatizer().lemmatize(text, pos='v')

def preprocess(text):
    result = []
    for token in gensim.utils.simple_preprocess(text):
        if token not in gensim.parsing.preprocessing.STOPWORDS and len(token) > 5:
            result.append(lemmatize(token))
    return result

# nltk.download('omw-1.4') # one-time download be run in case the library is not present on your local environment.

data['Title'] = data['Title'].map(preprocess)
data['Description'] = data['Description'].map(preprocess)

data.to_csv("dataset/preprocessed_data.csv")

## **Seperation of Master Reports and Duplicate Reports**

#Number of master_reports
data['Duplicated_issue'].isnull().sum()

#Saving all the duplicate reports into a csv file as a testing set
duplicate_reports = data.dropna(axis=0, subset=['Duplicated_issue'])
duplicate_reports.reset_index(drop=True)
duplicate_reports.to_csv('dataset/duplicate_reports.csv')

#Seperating all the master reports into a dataframe
master_reports = data[data.isnull().any(axis=1)]
master_reports.reset_index(drop=True)

#print('MASTER_REPORTS at INDEX #      ', master_reports[22])

# master_reports.to_csv(r'dataset\my_data_master.csv', index=False)

# master_reports['Issue_id'].replace('', np.nan, inplace=True)

# master_reports.dropna(subset=['Issue_id'], inplace=True)

# master_reports.to_csv(r'dataset\my_data_master_new.csv', index=False)

#from tabulate import tabulate
#display(master_reports.Description)
#print(tabulate(master_reports, headers = 'keys', tablefmt = 'psql'))


#dataframe = pd.DataFrame(master_reports, columns=master_reports.Description)
# printing data frame
#print(dataframe)



#print('LENGTH of MASTER_REPORTS      ', len(master_reports))
for i in range(len(master_reports)):
    print('INDEX      ', i)
    try:
        index = master_reports.Issue_id[i]
        print('Issue_id:           ', index)
        print('Description:           ', master_reports.Description[i])
    except KeyError as ke:
        print('Key Not Found in master_reports Dictionary:     ', ke)
        
    #if(i != 24):
    #print(master_reports.Duplicated_issue[i])

#Removing empty master_reports after data cleaning process
mr = pd.DataFrame()
count = 0
exception_count = 0
# print('LENGTH of MASTER_REPORTS      ', len(master_reports))
# print('')
# #print('MASTER_REPORTS at INDEX #      ', master_reports[2])
# print('')
# print('')
# print('')
for i in range(len(master_reports)):
    #print('COUNTER Value:         ', count)
    #print('Dictionary INDEX - I:     ', i)
    #print('master_reports Description:    ', master_reports.Description[i])
    
    try:
        if(len(master_reports.Description[i]) > 2):        
            #print('LEN of master_reports Description:    ', len(master_reports.Description[i]))
            mr = mr.append(master_reports.loc[[i]]) # use pandas.concat instead of pandas.append
            count+=1
    except KeyError as ke:
            #print('Key Not Found in master_reports Dictionary:     ', ke)
            exception_count+=1
        
        #print('REVISED MR     ', mr)
        

print('Total # Indices missing:       ', exception_count)

#Saving all the master reports into a seperate csv file
master_reports.to_csv('dataset/master_reports.csv')
